requirements.txt
streamlit
google-generativeai
python-dotenvg